+++ CHANGELOG.md
@@ -13,6 +13,10 @@
   for a in "123" as String? { echo $a; }
   ```
 
+#### Builtin
+
+- ``Error`` type constructor now accept `0` status
+
 ### Fixed
 
 - fix help message of ``prompt`` command in ``prompt`` module